pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "muppinllm"
version = "1.0.0"
description = "AI-Powered Crypto Market Analyst for Solana Tokens - Bullish or Bearish Analysis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Muppin Team", email = "team@muppin.fun"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["crypto", "solana", "ai", "llm", "market-analysis", "trading", "bullish", "bearish", "dex", "defi"]
dependencies = [
    "emergentintegrations>=0.1.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "isort>=5.13.0",
]

[project.urls]
Homepage = "https://muppin.fun"
Repository = "https://github.com/muppin/muppinllm"
Documentation = "https://github.com/muppin/muppinllm#readme"
Issues = "https://github.com/muppin/muppinllm/issues"

[project.scripts]
muppinllm = "muppinllm.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["muppinllm*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100
